refactor: move fromroe impl from repository to model
